Corrected LD bug
Originally, the limb darkening interpolation returned NaN when the star's parameters were outside the range covered by the input grid. Now, the closest grid point is used when input values fall outside the covered parameter range. This is important because the NaN limb darkening values caused the BATMAN model lightcurve to also be set to NaN during transit, which led to a NaN likelihood and prevented TATER from fitting the transit.